+++ lib/libesp32/Berry/default/be_tasmotalib.c
@@ -1412,7 +1412,7 @@ be_local_closure(gc,   /* name */
 ********************************************************************/
 be_local_closure(event,   /* name */
   be_nested_proto(
-    19,                          /* nstack */
+    20,                          /* nstack */
     6,                          /* argc */
     0,                          /* varg */
     0,                          /* has upvals */
@@ -1420,7 +1420,7 @@ be_local_closure(event,   /* name */
     0,                          /* has sup protos */
     NULL,                       /* no sub protos */
     1,                          /* has constants */
-    ( &(const bvalue[15]) {     /* constants */
+    ( &(const bvalue[18]) {     /* constants */
     /* K0   */  be_nested_string("introspect", 164638290, 10),
     /* K1   */  be_nested_string("every_50ms", -1911083288, 10),
     /* K2   */  be_nested_string("run_deferred", 371594696, 12),
@@ -1436,89 +1436,96 @@ be_local_closure(event,   /* name */
     /* K12  */  be_nested_string("format", -1180859054, 6),
     /* K13  */  be_nested_string("BRY: Exception> '%s' - %s", -2047976332, 25),
     /* K14  */  be_nested_string("stop_iteration", -121173395, 14),
+    /* K15  */  be_nested_string("save_before_restart", 1253239338, 19),
+    /* K16  */  be_nested_string("persist", -377883517, 7),
+    /* K17  */  be_nested_string("save", -855671224, 4),
     }),
     (be_nested_const_str("event", -30355297, 5)),
-    (be_nested_const_str("Tasmota.be", 825809411, 10)),
-    ( &(const binstruction[79]) {  /* code */
-      0xA41A0000,  //  0000  IMPORT R6  K0
-      0x1C1C0301,  //  0001  EQ R7  R1  K1
-      0x781E0001,  //  0002  JMPF R7  #0005
-      0x8C1C0102,  //  0003  GETMET R7  R0  K2
-      0x7C1C0200,  //  0004  CALL R7  1
-      0x1C1C0303,  //  0005  EQ R7  R1  K3
-      0x781E0006,  //  0006  JMPF R7  #000E
-      0x8C1C0104,  //  0007  GETMET R7  R0  K4
-      0x5C240400,  //  0008  MOVE R9  R2
-      0x5C280600,  //  0009  MOVE R10 R3
-      0x5C2C0800,  //  000A  MOVE R11 R4
-      0x7C1C0800,  //  000B  CALL R7  4
-      0x80040E00,  //  000C  RET  1 R7
-      0x7002003F,  //  000D  JMP    #004E
-      0x1C1C0305,  //  000E  EQ R7  R1  K5
-      0x781E0004,  //  000F  JMPF R7  #0015
-      0x8C1C0106,  //  0010  GETMET R7  R0  K6
-      0x5C240800,  //  0011  MOVE R9  R4
-      0x7C1C0400,  //  0012  CALL R7  2
-      0x80040E00,  //  0013  RET  1 R7
-      0x70020038,  //  0014  JMP    #004E
-      0x1C1C0307,  //  0015  EQ R7  R1  K7
-      0x781E0003,  //  0016  JMPF R7  #001B
-      0x8C1C0107,  //  0017  GETMET R7  R0  K7
-      0x7C1C0200,  //  0018  CALL R7  1
-      0x80040E00,  //  0019  RET  1 R7
-      0x70020032,  //  001A  JMP    #004E
-      0x881C0108,  //  001B  GETMBR R7  R0  K8
-      0x781E0030,  //  001C  JMPF R7  #004E
-      0x601C0010,  //  001D  GETGBL R7  G16
-      0x88200108,  //  001E  GETMBR R8  R0  K8
-      0x7C1C0200,  //  001F  CALL R7  1
-      0xA8020027,  //  0020  EXBLK  0 #0049
-      0x5C200E00,  //  0021  MOVE R8  R7
-      0x7C200000,  //  0022  CALL R8  0
-      0x8C240D09,  //  0023  GETMET R9  R6  K9
-      0x5C2C1000,  //  0024  MOVE R11 R8
-      0x5C300200,  //  0025  MOVE R12 R1
-      0x7C240600,  //  0026  CALL R9  3
-      0x60280004,  //  0027  GETGBL R10 G4
-      0x5C2C1200,  //  0028  MOVE R11 R9
-      0x7C280200,  //  0029  CALL R10 1
-      0x1C28150A,  //  002A  EQ R10 R10 K10
-      0x782A001B,  //  002B  JMPF R10 #0048
-      0xA802000E,  //  002C  EXBLK  0 #003C
-      0x5C281200,  //  002D  MOVE R10 R9
-      0x5C2C1000,  //  002E  MOVE R11 R8
-      0x5C300400,  //  002F  MOVE R12 R2
-      0x5C340600,  //  0030  MOVE R13 R3
-      0x5C380800,  //  0031  MOVE R14 R4
-      0x5C3C0A00,  //  0032  MOVE R15 R5
-      0x7C280A00,  //  0033  CALL R10 5
-      0x502C0200,  //  0034  LDBOOL R11 1 0
-      0x1C2C140B,  //  0035  EQ R11 R10 R11
-      0x782E0002,  //  0036  JMPF R11 #003A
-      0x502C0200,  //  0037  LDBOOL R11 1 0
-      0xA8040002,  //  0038  EXBLK  1 2
-      0x80041600,  //  0039  RET  1 R11
-      0xA8040001,  //  003A  EXBLK  1 1
-      0x7002000B,  //  003B  JMP    #0048
-      0xAC280002,  //  003C  CATCH  R10 0 2
-      0x70020008,  //  003D  JMP    #0047
-      0xA4321600,  //  003E  IMPORT R12 K11
-      0x60340001,  //  003F  GETGBL R13 G1
-      0x8C38190C,  //  0040  GETMET R14 R12 K12
-      0x5840000D,  //  0041  LDCONST  R16 K13
-      0x5C441400,  //  0042  MOVE R17 R10
-      0x5C481600,  //  0043  MOVE R18 R11
-      0x7C380800,  //  0044  CALL R14 4
-      0x7C340200,  //  0045  CALL R13 1
-      0x70020000,  //  0046  JMP    #0048
-      0xB0080000,  //  0047  RAISE  2 R0  R0
-      0x7001FFD7,  //  0048  JMP    #0021
-      0x581C000E,  //  0049  LDCONST  R7  K14
-      0xAC1C0200,  //  004A  CATCH  R7  1 0
-      0xB0080000,  //  004B  RAISE  2 R0  R0
-      0x501C0000,  //  004C  LDBOOL R7  0 0
-      0x80040E00,  //  004D  RET  1 R7
-      0x80000000,  //  004E  RET  0
+    (be_nested_const_str("input", -103256197, 5)),
+    ( &(const binstruction[83]) {  /* code */
+      0xA41A0000,  //  0000  IMPORT	R6	K0
+      0x1C1C0301,  //  0001  EQ	R7	R1	K1
+      0x781E0001,  //  0002  JMPF	R7	#0005
+      0x8C1C0102,  //  0003  GETMET	R7	R0	K2
+      0x7C1C0200,  //  0004  CALL	R7	1
+      0x501C0000,  //  0005  LDBOOL	R7	0	0
+      0x1C200303,  //  0006  EQ	R8	R1	K3
+      0x78220006,  //  0007  JMPF	R8	#000F
+      0x8C200104,  //  0008  GETMET	R8	R0	K4
+      0x5C280400,  //  0009  MOVE	R10	R2
+      0x5C2C0600,  //  000A  MOVE	R11	R3
+      0x5C300800,  //  000B  MOVE	R12	R4
+      0x7C200800,  //  000C  CALL	R8	4
+      0x80041000,  //  000D  RET	1	R8
+      0x7002003D,  //  000E  JMP		#004D
+      0x1C200305,  //  000F  EQ	R8	R1	K5
+      0x78220004,  //  0010  JMPF	R8	#0016
+      0x8C200106,  //  0011  GETMET	R8	R0	K6
+      0x5C280800,  //  0012  MOVE	R10	R4
+      0x7C200400,  //  0013  CALL	R8	2
+      0x80041000,  //  0014  RET	1	R8
+      0x70020036,  //  0015  JMP		#004D
+      0x1C200307,  //  0016  EQ	R8	R1	K7
+      0x78220003,  //  0017  JMPF	R8	#001C
+      0x8C200107,  //  0018  GETMET	R8	R0	K7
+      0x7C200200,  //  0019  CALL	R8	1
+      0x80041000,  //  001A  RET	1	R8
+      0x70020030,  //  001B  JMP		#004D
+      0x88200108,  //  001C  GETMBR	R8	R0	K8
+      0x7822002E,  //  001D  JMPF	R8	#004D
+      0x60200010,  //  001E  GETGBL	R8	G16
+      0x88240108,  //  001F  GETMBR	R9	R0	K8
+      0x7C200200,  //  0020  CALL	R8	1
+      0xA8020027,  //  0021  EXBLK	0	#004A
+      0x5C241000,  //  0022  MOVE	R9	R8
+      0x7C240000,  //  0023  CALL	R9	0
+      0x8C280D09,  //  0024  GETMET	R10	R6	K9
+      0x5C301200,  //  0025  MOVE	R12	R9
+      0x5C340200,  //  0026  MOVE	R13	R1
+      0x7C280600,  //  0027  CALL	R10	3
+      0x602C0004,  //  0028  GETGBL	R11	G4
+      0x5C301400,  //  0029  MOVE	R12	R10
+      0x7C2C0200,  //  002A  CALL	R11	1
+      0x1C2C170A,  //  002B  EQ	R11	R11	K10
+      0x782E0019,  //  002C  JMPF	R11	#0047
+      0xA802000C,  //  002D  EXBLK	0	#003B
+      0x5C2C1400,  //  002E  MOVE	R11	R10
+      0x5C301200,  //  002F  MOVE	R12	R9
+      0x5C340400,  //  0030  MOVE	R13	R2
+      0x5C380600,  //  0031  MOVE	R14	R3
+      0x5C3C0800,  //  0032  MOVE	R15	R4
+      0x5C400A00,  //  0033  MOVE	R16	R5
+      0x7C2C0A00,  //  0034  CALL	R11	5
+      0x5C1C1600,  //  0035  MOVE	R7	R11
+      0x781E0001,  //  0036  JMPF	R7	#0039
+      0xA8040001,  //  0037  EXBLK	1	1
+      0x7002000E,  //  0038  JMP		#0048
+      0xA8040001,  //  0039  EXBLK	1	1
+      0x7002000B,  //  003A  JMP		#0047
+      0xAC2C0002,  //  003B  CATCH	R11	0	2
+      0x70020008,  //  003C  JMP		#0046
+      0xA4361600,  //  003D  IMPORT	R13	K11
+      0x60380001,  //  003E  GETGBL	R14	G1
+      0x8C3C1B0C,  //  003F  GETMET	R15	R13	K12
+      0x5844000D,  //  0040  LDCONST	R17	K13
+      0x5C481600,  //  0041  MOVE	R18	R11
+      0x5C4C1800,  //  0042  MOVE	R19	R12
+      0x7C3C0800,  //  0043  CALL	R15	4
+      0x7C380200,  //  0044  CALL	R14	1
+      0x70020000,  //  0045  JMP		#0047
+      0xB0080000,  //  0046  RAISE	2	R0	R0
+      0x7001FFD9,  //  0047  JMP		#0022
+      0xA8040001,  //  0048  EXBLK	1	1
+      0x70020002,  //  0049  JMP		#004D
+      0x5820000E,  //  004A  LDCONST	R8	K14
+      0xAC200200,  //  004B  CATCH	R8	1	0
+      0xB0080000,  //  004C  RAISE	2	R0	R0
+      0x1C20030F,  //  004D  EQ	R8	R1	K15
+      0x78220002,  //  004E  JMPF	R8	#0052
+      0xA4222000,  //  004F  IMPORT	R8	K16
+      0x8C241111,  //  0050  GETMET	R9	R8	K17
+      0x7C240200,  //  0051  CALL	R9	1
+      0x80040E00,  //  0052  RET	1	R7
     })
   )
 );